requirements/dev.txt
-r ./base.txt
-e git://github.com/domain51/d51.django.virtualenv.base#egg=d51.django.virtualenv.base
-e git://github.com/domain51/d51.django.virtualenv.test_runner#egg=d51.django.virtualenv.test_runner
armstrong.dev

# This is needed to avoid a hang in case a Django server is already running
-e git://github.com/tswicegood/lettuce.git@fix-port-issues#egg=lettuce
